refactor: Move encipherment modules to src directory
Renamed cipher.py and homophones.py from encipherment/ to src/encipherment/. Updated import in cipher.py to use relative import for extract_homophones.
