+++ src/encipherment/cipher.py
@@ -1,5 +1,5 @@
 from decimal import Decimal
-from homophones import extract_homophones
+from .homophones import extract_homophones
 import random
 
 class Cipher:
